refactor: user를 member로 변경
