+++ back/src/main/java/com/baba/back/oauth/OAuthClient.java
@@ -6,5 +6,5 @@ public interface OAuthClient {
 
     OAuthAccessTokenResponse getOAuthAccessToken(String code);
 
-    String getUserId(String accessToken);
+    String getMemberId(String accessToken);
 }
