+++ back/src/main/java/com/baba/back/oauth/repository/JoinedUserRepository.java
@@ -1,8 +1,8 @@
 package com.baba.back.oauth.repository;
 
-import com.baba.back.oauth.domain.JoinedUser;
+import com.baba.back.oauth.domain.JoinedMember;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface JoinedUserRepository extends JpaRepository<JoinedUser, String> {
+public interface JoinedUserRepository extends JpaRepository<JoinedMember, String> {
 
 }
