+++ back/src/main/java/com/baba/back/oauth/service/OAuthService.java
@@ -1,7 +1,7 @@
 package com.baba.back.oauth.service;
 
 import com.baba.back.oauth.OAuthClient;
-import com.baba.back.oauth.domain.JoinedUser;
+import com.baba.back.oauth.domain.JoinedMember;
 import com.baba.back.oauth.dto.OAuthAccessTokenResponse;
 import com.baba.back.oauth.dto.TokenResponse;
 import com.baba.back.oauth.repository.JoinedUserRepository;
@@ -22,12 +22,12 @@ public class OAuthService {
 
     public TokenResponse signInKakao(final String code) {
         final OAuthAccessTokenResponse accessTokenResponse = oAuthClient.getOAuthAccessToken(code);
-        final String userId = oAuthClient.getUserId(accessTokenResponse.getAccessToken());
-        final JoinedUser joinedUser = joinedUserRepository.findById(userId)
-                .orElseGet(() -> joinedUserRepository.save(new JoinedUser(userId, false)));
-        final String token = tokenProvider.createToken(userId);
+        final String memberId = oAuthClient.getMemberId(accessTokenResponse.getAccessToken());
+        final JoinedMember joinedMember = joinedUserRepository.findById(memberId)
+                .orElseGet(() -> joinedUserRepository.save(new JoinedMember(memberId, false)));
+        final String token = tokenProvider.createToken(memberId);
 
-        if (joinedUser.getIsSigned()) {
+        if (joinedMember.isSigned()) {
             return new TokenResponse(true, "이미 가입되어 있습니다.", token);
         }
         return new TokenResponse(false, "가입이 필요합니다.", token);
