+++ back/src/test/java/com/baba/back/oauth/domain/member/IntroductionTest.java
@@ -1,4 +1,4 @@
-package com.baba.back.oauth.domain.user;
+package com.baba.back.oauth.domain.member;
 
 
 import com.baba.back.oauth.exception.IntroductionLengthBadRequestException;
