+++ back/src/test/java/com/baba/back/oauth/domain/member/NameTest.java
@@ -1,4 +1,4 @@
-package com.baba.back.oauth.domain.user;
+package com.baba.back.oauth.domain.member;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
